backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
Pillow==10.1.0
firebase-admin==6.2.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1